07-ai-powered-pipeline-corrosion-analysis/src/cicd-pipeline-code/continuous-integration-delivery/requirements.txt
scikit-learn==1.3.0
pandas==1.5.0
joblib==1.3.2
numpy==1.23.3
ai-core-sdk==1.22.3
pandas-schema==0.3.6
requests==2.31.0
ipython==8.15.0
pyyaml==6.0.1
glob2==0.7
GitPython==3.1.37
docker==6.1.3
Flask==3.0.0
gunicorn==21.2.0